+++ inst/dependencies.R
@@ -1,2 +1,0 @@
-pkgnet::CreatePackageReport("citeme", report_path = "inst/report.html")
-covr::report()